// libs/api-client/model/department.ts
/**
 * API IP32
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Team } from './team';


export interface Department { 
    derpartmentID?: number;
    name: string;
    teams?: Array<Team> | null;
}